requirements.txt
apache-beam==2.29.0
avro-python3==1.9.2.1
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
brotlipy==0.7.0
cachetools==4.2.2
certifi==2020.12.5
cffi @ file:///C:/ci/cffi_1606237165371/work
chardet @ file:///C:/ci/chardet_1605296577955/work
cloudpickle @ file:///tmp/build/80754af9/cloudpickle_1598884132938/work
colorama @ file:///tmp/build/80754af9/colorama_1607707115595/work
conda==4.10.1
conda-package-handling @ file:///C:/ci/conda-package-handling_1605484860147/work
crcmod==1.7
cryptography @ file:///C:/ci/cryptography_1607618721363/work
DateTime==4.3
decorator @ file:///tmp/build/80754af9/decorator_1621259047763/work
dill==0.3.1.1
docopt==0.6.2
facets-overview==1.0.0
fastavro==1.4.1
fasteners==0.16
future==0.18.2
google-api-core==1.28.0
google-apitools==0.5.31
google-auth==1.30.0
google-cloud==0.34.0
google-cloud-bigquery==1.28.0
google-cloud-bigtable==1.7.0
google-cloud-core==1.6.0
google-cloud-datastore==1.15.3
google-cloud-dlp==1.0.0
google-cloud-language==1.3.0
google-cloud-pubsub==1.7.0
google-cloud-spanner==1.19.1
google-cloud-storage==1.38.0
google-cloud-videointelligence==1.16.1
google-cloud-vision==1.0.0
google-crc32c==1.1.2
google-resumable-media==1.3.0
googleapis-common-protos==1.53.0
grpc-google-iam-v1==0.12.3
grpcio==1.38.0
grpcio-gcp==0.2.2
hdfs==2.6.0
httplib2==0.17.4
idna @ file:///tmp/build/80754af9/idna_1593446292537/work
ipykernel @ file:///C:/ci/ipykernel_1607454116140/work/dist/ipykernel-5.3.4-py3-none-any.whl
ipython @ file:///C:/ci/ipython_1617120911285/work
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
jedi @ file:///C:/ci/jedi_1606932640618/work
josepy==1.8.0
jupyter-client @ file:///tmp/build/80754af9/jupyter_client_1616770841739/work
jupyter-core @ file:///C:/ci/jupyter_core_1612213515523/work
language==0.6
libcst==0.3.19
menuinst==1.4.16
mypy-extensions==0.4.3
numpy==1.20.3
oauth2client==4.1.3
packaging==20.9
pandas==1.2.4
parso==0.7.0
pickle-mixin==1.0.2
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1616415428029/work
proto-plus==1.18.1
protobuf==3.17.0
pyarrow==3.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycosat==0.6.3
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
pydot==1.4.2
Pygments @ file:///tmp/build/80754af9/pygments_1621606182707/work
pymongo==3.11.4
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1606517880428/work
pyparsing==2.4.7
pyRFC3339==1.1
PySocks @ file:///C:/ci/pysocks_1605307512533/work
python-dateutil @ file:///home/ktietz/src/ci/python-dateutil_1611928101742/work
pytz==2021.1
pywin32==228
PyYAML==5.4.1
pyzmq==20.0.0
requests @ file:///tmp/build/80754af9/requests_1606691187061/work
requests-toolbelt==0.9.1
rsa==4.7.2
ruamel-yaml-conda @ file:///C:/ci/ruamel_yaml_1605524207026/work
six @ file:///C:/ci/six_1605205404068/work
spyder-kernels @ file:///C:/ci/spyder-kernels_1617396763091/work
timeloop==1.0.2
tornado @ file:///C:/ci/tornado_1606924294691/work
tqdm @ file:///tmp/build/80754af9/tqdm_1607369919789/work
traitlets @ file:///home/ktietz/src/ci/traitlets_1611929699868/work
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3 @ file:///tmp/build/80754af9/urllib3_1603305693037/work
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306162074/work
wincertstore==0.2
zope.component==5.0.0
zope.event==4.5.0
zope.hookable==5.0.1
zope.interface==5.4.0
